Add ArgParse for initialization of gnosis_cli.py
Add ArgParse for initialization of gnosis_cli.py with Exception control when ganache-cli it's not running.
Add new methods to NetworkAgent, now this module manages the change of EthereumClient in the console
Add Logger coverage in the console_getter and console_engine to properly track the activation of the commands.